+++ Testing Code/I2C Test/I2C_Test_Master/feedback.h
@@ -2,7 +2,9 @@
 #define _FEEDBACK_H
 
 #include "helper.h"
-// Define global variables for panel feedback 
+#include <Wire.h>
+
+// Define global variables for panel feedback
 extern volatile bool panelStatus;  // Retracted
 
 // Define global variables for simple collection feedback
@@ -14,9 +16,9 @@ extern volatile float currentPowerDraw;        // in Amps
 extern volatile float currentPowerLevel;       // in percentage
 extern volatile float currentCoreTemperature;  // in Celsius
 
-// Poll status form the panel deployment system
+// Poll status from the panel deployment system
 void requestPanelState(byte slaveAddr) {
-  Wire.requestFrom(slaveAddr, 1);  // Request 1 byte from the slave
+  Wire.requestFrom((uint8_t)slaveAddr, (uint8_t)1);  // Request 1 byte from the slave
   if (Wire.available()) {
     panelStatus = Wire.read();
   }
@@ -24,7 +26,7 @@ void requestPanelState(byte slaveAddr) {
 
 // Poll status from sample collection system
 void requestSamplingState(byte slaveAddr) {
-  Wire.requestFrom(slaveAddr, 2);  // Request 2 bytes from the slave
+  Wire.requestFrom((uint8_t)slaveAddr, (uint8_t)2);  // Request 2 bytes from the slave
   if (Wire.available() == 2) {
     byte DrillState = Wire.read();
     byte ArmState = Wire.read();
@@ -35,11 +37,8 @@ void requestSamplingState(byte slaveAddr) {
 }
 
 void requestBatteryState(byte slaveAddr) {
-  Wire.requestFrom(slaveAddr, 12);  // Request 12 bytes from the slave
-  delay(10);                        // Small delay to ensure data is available
-
-  // Serial.print("Bytes available: ");
-  // Serial.println(Wire.available());
+  Wire.requestFrom((uint8_t)slaveAddr, (uint8_t)12);  // Request 12 bytes from the slave
+  delay(10);                                          // Small delay to ensure data is available
 
   if (Wire.available() == 12) {
     byte rawData[12];
